+++ src/renderer_raster_clusters_lod.cpp
@@ -435,13 +435,7 @@ void RendererRasterClustersLod::render(VkCommandBuffer cmd, Resources& res, Rend
 {
   VkMemoryBarrier memBarrier = {VK_STRUCTURE_TYPE_MEMORY_BARRIER};
 
-  {
-    glm::vec2 renderScale = res.getFramebufferWindow2RenderScale();
-    float     pixelScale  = std::min(renderScale.x, renderScale.y);
-
-    m_sceneBuildShaderio.errorOverDistanceThreshold =
-        clusterLodErrorOverDistance(frame.lodPixelError * pixelScale, frame.traversalFov, frame.traversalViewHeight);
-  }
+  m_sceneBuildShaderio.errorOverDistanceThreshold = updateLodPixelError(res, rscene, frame);
 
   m_sceneBuildShaderio.traversalViewMatrix    = frame.traversalViewMatrix;
   m_sceneBuildShaderio.cullViewProjMatrix     = frame.cullViewProjMatrix;